refactor(rules): split di_hygiene.go and observability.go by domain
Mechanical split — no logic, behavior, or naming changes. Rule types,
methods, and exported names are preserved exactly; tests and registry
files reference rule types by name within the same package.

di_hygiene.go (1809 lines) split into:
- di_anvil.go         — AnvilMergeComponentEmptyScopeRule and
                        AnvilContributesBindingWithoutScopeRule
- di_cycles.go        — DiCycleDetectionRule and formatDICycle
- di_duplication.go   — IntoMapDuplicateKeyRule, IntoSetDuplicateTypeRule
                        and their callee helpers
- di_hilt.go          — HiltEntryPointOnNonInterfaceRule,
                        HiltInstallInMismatchRule,
                        HiltSingletonWithActivityDepRule and their tables
- di_hygiene.go       — remaining DI rules (Binds*, DeadBindings, Provider/
                        Lazy, Subcomponent, ComponentMissingModule, Metro,
                        Scope, JvmSuppressWildcards, IntoMap/IntoSet missing/
                        wrapper, SingletonOnMutableClass) plus shared
                        annotation/type/import helpers reused by these
                        rules, database.go, and registry_di_hygiene.go.

observability.go (2035 lines) split into:
- logging.go   — LogLevelGuardMissing, LogWithoutCorrelationID, Logger*,
                 Unstructured/TraceID/StructuredLogKey, LoggerStringConcat,
                 Mdc* rules and NullableStructuredField, plus their
                 rule-private helpers.
- tracing.go   — WithContextWithoutTracingContext, SpanStartWithoutFinish,
                 SpanAttributeWithHighCardinality.
- metrics.go   — MetricTimerOutsideBlock, MetricTagHighCardinality,
                 MetricNameMissingUnit, MetricCounterNotMonotonic plus
                 metricConstructorNames.
- observability.go — shared helpers only: ifConditionThenElseBodiesFlat
                     (also used by i18n_plurals), coroutineBuilderPartsFlat
                     and coroutineContextHasMDCFlat (logging + tracing),
                     firstHighCardinalityString (tracing + metrics), and
                     the throwableLikeIdentifier / interpolationIdentifier
                     regexes (logging rules + observability_helpers).

Validation: go build, go vet, golangci-lint run ./..., and go test
./internal/rules/ -count=1 all pass.
