+++ internal/rules/di_hygiene.go
@@ -5,332 +5,11 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/kaeawc/krit/internal/di"
 	api "github.com/kaeawc/krit/internal/rules/api"
 	"github.com/kaeawc/krit/internal/scanner"
 )
 
 const diHygieneRuleSet = "di-hygiene"
-
-// AnvilMergeComponentEmptyScopeRule detects @MergeComponent scopes that have no
-// matching @ContributesTo/@ContributesBinding declarations anywhere in the project.
-type AnvilMergeComponentEmptyScopeRule struct {
-	FlatDispatchBase
-	BaseRule
-}
-
-type DiCycleDetectionRule struct {
-	FlatDispatchBase
-	BaseRule
-}
-
-// Confidence reports a tier-2 (medium) base confidence. DI hygiene rule. Detection uses annotation and import patterns for
-// Dagger/Hilt/Anvil; project-specific DI aliases are not followed.
-// Classified per roadmap/17.
-func (r *AnvilMergeComponentEmptyScopeRule) Confidence() float64 { return 0.75 }
-
-func (r *DiCycleDetectionRule) Confidence() float64 { return 0.75 }
-
-func (r *DiCycleDetectionRule) check(ctx *api.Context) {
-	if len(ctx.ParsedFiles) == 0 {
-		return
-	}
-	graph := di.BuildGraph(ctx.ParsedFiles, nil)
-	for _, cycle := range graph.FindCycles() {
-		if len(cycle.Bindings) == 0 || cycle.Bindings[0] == nil {
-			continue
-		}
-		first := cycle.Bindings[0]
-		ctx.Emit(scanner.Finding{
-			File:     first.File,
-			Line:     first.Line,
-			Col:      1,
-			RuleSet:  r.RuleSetName,
-			Rule:     r.RuleName,
-			Severity: r.Sev,
-			Message:  fmt.Sprintf("DI binding cycle detected: %s.", formatDICycle(cycle)),
-		})
-	}
-}
-
-func formatDICycle(cycle di.Cycle) string {
-	names := make([]string, 0, len(cycle.Bindings)+1)
-	for _, b := range cycle.Bindings {
-		if b != nil {
-			names = append(names, b.FQN)
-		}
-	}
-	if len(names) > 0 {
-		names = append(names, names[0])
-	}
-	return strings.Join(names, " -> ")
-}
-
-func (r *AnvilMergeComponentEmptyScopeRule) check(ctx *api.Context) {
-	index := ctx.CodeIndex
-	if index == nil {
-		return
-	}
-	files := index.Files
-	if len(files) == 0 {
-		return
-	}
-
-	contributedScopes := make(map[string]struct{})
-	type mergeComponentCandidate struct {
-		file  *scanner.File
-		idx   uint32
-		scope string
-	}
-	var mergeComponents []mergeComponentCandidate
-
-	for _, file := range files {
-		if file == nil || file.FlatTree == nil || !anvilMergeComponentMayMatch(file.Content) {
-			continue
-		}
-		for _, idx := range anvilScopeDeclarationCandidates(file) {
-			if !anvilModifiersMayMatchFlat(file, idx) {
-				continue
-			}
-
-			if scope := anvilAnnotationScopeFlat(file, idx, "ContributesTo"); scope != "" {
-				contributedScopes[scope] = struct{}{}
-			}
-			if scope := anvilAnnotationScopeFlat(file, idx, "ContributesBinding"); scope != "" {
-				contributedScopes[scope] = struct{}{}
-			}
-			if scope := anvilAnnotationScopeFlat(file, idx, "MergeComponent"); scope != "" {
-				mergeComponents = append(mergeComponents, mergeComponentCandidate{
-					file:  file,
-					idx:   idx,
-					scope: scope,
-				})
-			}
-		}
-	}
-
-	for _, candidate := range mergeComponents {
-		if _, ok := contributedScopes[candidate.scope]; ok {
-			continue
-		}
-
-		name := extractIdentifierFlat(candidate.file, candidate.idx)
-		if name == "" {
-			name = "merged component"
-		}
-
-		ctx.Emit(r.Finding(
-			candidate.file,
-			candidate.file.FlatRow(candidate.idx)+1,
-			1,
-			fmt.Sprintf("@MergeComponent(%s::class) on '%s' has no matching @ContributesTo or @ContributesBinding scope in the project, so the merged component will be empty.", candidate.scope, name),
-		))
-	}
-}
-
-func (r *IntoMapDuplicateKeyRule) check(ctx *api.Context) {
-	index := ctx.CodeIndex
-	if index == nil || len(index.Files) == 0 {
-		return
-	}
-	type entry struct {
-		file *scanner.File
-		idx  uint32
-		name string
-	}
-	groups := make(map[string][]entry)
-	for _, file := range index.Files {
-		if file == nil || file.FlatTree == nil {
-			continue
-		}
-		if !bytes.Contains(file.Content, []byte("IntoMap")) {
-			continue
-		}
-		pkg := sourcePackageName(file)
-		file.FlatWalkAllNodes(0, func(idx uint32) {
-			if file.FlatType(idx) != "function_declaration" {
-				return
-			}
-			if !hasAnnotationFlat(file, idx, "IntoMap") {
-				return
-			}
-			if !hasAnnotationFlat(file, idx, "Provides") && !hasAnnotationFlat(file, idx, "Binds") {
-				return
-			}
-			keyLit, ok := firstMapKeyAnnotationLiteralFlat(file, idx)
-			if !ok {
-				return
-			}
-			scope := pkg + "::" + enclosingClassChainFlat(file, idx)
-			groupKey := scope + "::" + keyLit
-			name := extractIdentifierFlat(file, idx)
-			groups[groupKey] = append(groups[groupKey], entry{file: file, idx: idx, name: name})
-		})
-	}
-	for groupKey, entries := range groups {
-		if len(entries) < 2 {
-			continue
-		}
-		keyLit := suffixAfterLastDoubleColon(groupKey)
-		for _, e := range entries {
-			n := e.name
-			if n == "" {
-				n = "binding"
-			}
-			ctx.Emit(r.Finding(
-				e.file,
-				e.file.FlatRow(e.idx)+1,
-				1,
-				fmt.Sprintf("@IntoMap function '%s' shares key %s with %d other contribution(s) in the same module/component; use a distinct map key or remove the duplicate contribution.", n, keyLit, len(entries)-1),
-			))
-		}
-	}
-}
-
-func suffixAfterLastDoubleColon(s string) string {
-	idx := strings.LastIndex(s, "::")
-	if idx < 0 {
-		return ""
-	}
-	return s[idx+2:]
-}
-
-func (r *IntoSetDuplicateTypeRule) check(ctx *api.Context) {
-	index := ctx.CodeIndex
-	if index == nil || len(index.Files) == 0 {
-		return
-	}
-	type entry struct {
-		file *scanner.File
-		idx  uint32
-		name string
-	}
-	groups := make(map[string][]entry)
-	for _, file := range index.Files {
-		if file == nil || file.FlatTree == nil {
-			continue
-		}
-		if !bytes.Contains(file.Content, []byte("IntoSet")) {
-			continue
-		}
-		pkg := sourcePackageName(file)
-		file.FlatWalkAllNodes(0, func(idx uint32) {
-			if file.FlatType(idx) != "function_declaration" {
-				return
-			}
-			if !hasAnnotationFlat(file, idx, "IntoSet") {
-				return
-			}
-			if !hasAnnotationFlat(file, idx, "Provides") {
-				return
-			}
-			impl := intoSetReturnedExpressionCallee(file, idx)
-			if impl == "" {
-				return
-			}
-			scope := pkg + "::" + enclosingClassChainFlat(file, idx)
-			groupKey := scope + "::" + impl
-			name := extractIdentifierFlat(file, idx)
-			groups[groupKey] = append(groups[groupKey], entry{file: file, idx: idx, name: name})
-		})
-	}
-	for groupKey, entries := range groups {
-		if len(entries) < 2 {
-			continue
-		}
-		impl := suffixAfterLastDoubleColon(groupKey)
-		for _, e := range entries {
-			n := e.name
-			if n == "" {
-				n = "binding"
-			}
-			ctx.Emit(r.Finding(
-				e.file,
-				e.file.FlatRow(e.idx)+1,
-				1,
-				fmt.Sprintf("@IntoSet function '%s' contributes '%s' which is also contributed by %d other binding(s) in the same module/component; the set dedupes, so contributions are lost.", n, impl, len(entries)-1),
-			))
-		}
-	}
-}
-
-// intoSetReturnedExpressionCallee returns the simple-identifier callee of the
-// expression-body or single-return statement of `idx`. Used to compare two
-// `@IntoSet` providers that construct the same concrete implementation.
-func intoSetReturnedExpressionCallee(file *scanner.File, idx uint32) string {
-	if file == nil || idx == 0 {
-		return ""
-	}
-	// Expression body: `= Foo()` — find call_expression sibling of `=`.
-	for child := file.FlatFirstChild(idx); child != 0; child = file.FlatNextSib(child) {
-		if file.FlatType(child) != "=" {
-			continue
-		}
-		next := file.FlatNextSib(child)
-		for next != 0 && !file.FlatIsNamed(next) {
-			next = file.FlatNextSib(next)
-		}
-		if next == 0 {
-			break
-		}
-		return calleeIdentifierText(file, next)
-	}
-	// Block body: look for the single return statement.
-	body, ok := file.FlatFindChild(idx, "function_body")
-	if !ok || body == 0 {
-		return ""
-	}
-	var lone uint32
-	for child := file.FlatFirstChild(body); child != 0; child = file.FlatNextSib(child) {
-		if !file.FlatIsNamed(child) {
-			continue
-		}
-		t := file.FlatType(child)
-		if t == "{" || t == "}" {
-			continue
-		}
-		if lone != 0 {
-			return ""
-		}
-		lone = child
-	}
-	if lone == 0 {
-		return ""
-	}
-	if file.FlatType(lone) == "jump_expression" {
-		// `return Foo()`: walk down to call_expression
-		for c := file.FlatFirstChild(lone); c != 0; c = file.FlatNextSib(c) {
-			if file.FlatIsNamed(c) {
-				return calleeIdentifierText(file, c)
-			}
-		}
-	}
-	return calleeIdentifierText(file, lone)
-}
-
-func calleeIdentifierText(file *scanner.File, idx uint32) string {
-	if file.FlatType(idx) != "call_expression" {
-		return ""
-	}
-	callee := file.FlatFirstChild(idx)
-	for callee != 0 && !file.FlatIsNamed(callee) {
-		callee = file.FlatNextSib(callee)
-	}
-	if callee == 0 {
-		return ""
-	}
-	switch file.FlatType(callee) {
-	case "simple_identifier":
-		return strings.TrimSpace(file.FlatNodeText(callee))
-	case "navigation_expression":
-		text := strings.TrimSpace(file.FlatNodeText(callee))
-		if dot := strings.LastIndex(text, "."); dot >= 0 {
-			return text[dot+1:]
-		}
-		return text
-	}
-	return ""
-}
 
 var (
 	anvilContributesToToken      = []byte("ContributesTo")
@@ -375,18 +54,6 @@ func anvilModifiersMayMatchFlat(file *scanner.File, idx uint32) bool {
 		bytes.Contains(segment, anvilMergeComponentToken)
 }
 
-// AnvilContributesBindingWithoutScopeRule detects a same-file mismatch between
-// @ContributesBinding(scope) and the @ContributesTo(scope) on the bound interface.
-type AnvilContributesBindingWithoutScopeRule struct {
-	FlatDispatchBase
-	BaseRule
-}
-
-// Confidence reports a tier-2 (medium) base confidence. DI hygiene rule. Detection uses annotation and import patterns for
-// Dagger/Hilt/Anvil; project-specific DI aliases are not followed.
-// Classified per roadmap/17.
-func (r *AnvilContributesBindingWithoutScopeRule) Confidence() float64 { return 0.75 }
-
 // BindsMismatchedArityRule detects @Binds functions that do not declare exactly
 // one parameter, which Dagger rejects during code generation.
 type BindsMismatchedArityRule struct {
@@ -398,81 +65,6 @@ type BindsMismatchedArityRule struct {
 // Dagger/Hilt/Anvil; project-specific DI aliases are not followed.
 // Classified per roadmap/17.
 func (r *BindsMismatchedArityRule) Confidence() float64 { return 0.75 }
-
-// HiltEntryPointOnNonInterfaceRule detects Hilt entry points declared as a
-// class or object instead of an interface.
-type HiltEntryPointOnNonInterfaceRule struct {
-	FlatDispatchBase
-	BaseRule
-}
-
-// Confidence reports a tier-2 (medium) base confidence. DI hygiene rule. Detection uses annotation and import patterns for
-// Dagger/Hilt/Anvil; project-specific DI aliases are not followed.
-// Classified per roadmap/17.
-func (r *HiltEntryPointOnNonInterfaceRule) Confidence() float64 { return 0.75 }
-
-func hiltEntryPointDeclarationFlat(file *scanner.File, idx uint32) (kind, name string, line int, ok bool) {
-	if file == nil || !hasAnnotationFlat(file, idx, "EntryPoint") {
-		return "", "", 0, false
-	}
-
-	switch file.FlatType(idx) {
-	case "class_declaration", "object_declaration":
-		return hiltEntryPointDeclKindFlat(file, idx), extractIdentifierFlat(file, idx), file.FlatRow(idx) + 1, true
-	case "prefix_expression":
-		target := hiltEntryPointAnnotatedTargetFlat(file, idx)
-		if target == 0 {
-			return "", "", 0, false
-		}
-		switch file.FlatType(target) {
-		case "class_declaration", "object_declaration":
-			return hiltEntryPointDeclKindFlat(file, target), extractIdentifierFlat(file, target), file.FlatRow(idx) + 1, true
-		case "infix_expression":
-			return hiltEntryPointInfixDeclFlat(file, target)
-		}
-	}
-
-	return "", "", 0, false
-}
-
-func hiltEntryPointAnnotatedTargetFlat(file *scanner.File, idx uint32) uint32 {
-	current := idx
-	for file != nil && file.FlatType(current) == "prefix_expression" {
-		if file.FlatNamedChildCount(current) < 2 {
-			return 0
-		}
-		current = file.FlatNamedChild(current, 1)
-	}
-	return current
-}
-
-func hiltEntryPointInfixDeclFlat(file *scanner.File, idx uint32) (kind, name string, line int, ok bool) {
-	if file == nil || file.FlatNamedChildCount(idx) < 2 {
-		return "", "", 0, false
-	}
-
-	kind = file.FlatNodeText(file.FlatNamedChild(idx, 0))
-	if kind != "class" && kind != "interface" {
-		return "", "", 0, false
-	}
-
-	name = file.FlatNodeText(file.FlatNamedChild(idx, 1))
-	return kind, name, file.FlatRow(idx) + 1, true
-}
-
-func hiltEntryPointDeclKindFlat(file *scanner.File, idx uint32) string {
-	switch file.FlatType(idx) {
-	case "object_declaration":
-		return "object"
-	case "class_declaration":
-		if file.FlatHasChildOfType(idx, "interface") {
-			return "interface"
-		}
-		return "class"
-	default:
-		return "class"
-	}
-}
 
 func anvilContributedInterfaceScopesFlat(file *scanner.File) map[string]string {
 	scopes := make(map[string]string)
@@ -789,42 +381,6 @@ func isInsideComponentInterfaceFlat(file *scanner.File, idx uint32) bool {
 		}
 	}
 	return false
-}
-
-// HiltInstallInMismatchRule detects @Module/@InstallIn classes whose
-// `@Provides` functions are annotated with a Hilt scope that does not match
-// the module's installed component.
-type HiltInstallInMismatchRule struct {
-	FlatDispatchBase
-	BaseRule
-}
-
-// Confidence reports a tier-2 (medium) base confidence. DI hygiene rule.
-func (r *HiltInstallInMismatchRule) Confidence() float64 { return 0.75 }
-
-// hiltComponentScopes maps Hilt component class names to the scope annotations
-// they own. The scopes are stored as a set for fast membership checks.
-var hiltComponentScopes = map[string]map[string]struct{}{
-	"SingletonComponent":        {"Singleton": {}},
-	"ActivityRetainedComponent": {"ActivityRetainedScoped": {}},
-	"ActivityComponent":         {"ActivityScoped": {}},
-	"FragmentComponent":         {"FragmentScoped": {}},
-	"ViewComponent":             {"ViewScoped": {}},
-	"ViewWithFragmentComponent": {"ViewScoped": {}},
-	"ServiceComponent":          {"ServiceScoped": {}},
-	"ViewModelComponent":        {"ViewModelScoped": {}},
-}
-
-// hiltScopeAnnotations enumerates the Hilt scope annotation names. A function
-// annotated with one of these but a non-matching component is suspect.
-var hiltScopeAnnotations = []string{
-	"Singleton",
-	"ActivityRetainedScoped",
-	"ActivityScoped",
-	"FragmentScoped",
-	"ViewScoped",
-	"ServiceScoped",
-	"ViewModelScoped",
 }
 
 // InjectOnAbstractClassRule detects @Inject primary constructors on abstract
@@ -1382,30 +938,6 @@ func hasAnyInjectPrimaryConstructor(file *scanner.File, idx uint32) bool {
 	return hasAnnotationFlat(file, ctor, "Inject")
 }
 
-// IntoMapDuplicateKeyRule detects two or more @IntoMap @Provides/@Binds
-// functions in the same enclosing module/component that declare the same key
-// literal. Duplicate map keys create conflicting contributions.
-type IntoMapDuplicateKeyRule struct {
-	FlatDispatchBase
-	BaseRule
-}
-
-// Confidence reports a tier-2 (medium) base confidence. DI hygiene rule.
-func (r *IntoMapDuplicateKeyRule) Confidence() float64 { return 0.7 }
-
-// IntoSetDuplicateTypeRule detects two or more @IntoSet @Provides functions in
-// the same enclosing module/component that return the same concrete impl
-// expression — Dagger's set dedupes by value, so contributions are lost.
-type IntoSetDuplicateTypeRule struct {
-	FlatDispatchBase
-	BaseRule
-}
-
-// Confidence reports a tier-3 (lower) base confidence. Detection compares the
-// final returned-expression callee text; aliases and indirection are not
-// followed.
-func (r *IntoSetDuplicateTypeRule) Confidence() float64 { return 0.5 }
-
 // firstMapKeyAnnotationLiteralFlat returns the literal argument of the first
 // `@*Key(...)` annotation on idx (e.g. `"foo"` for `@StringKey("foo")`).
 // Returns ("", false) if no key annotation or no parenthesized argument.
@@ -1479,30 +1011,6 @@ func enclosingClassChainFlat(file *scanner.File, idx uint32) string {
 		}
 	}
 	return strings.Join(names, ".")
-}
-
-// HiltSingletonWithActivityDepRule detects @Singleton classes whose constructor
-// takes an Activity-, Fragment-, View-, or LifecycleOwner-scoped parameter,
-// which is a scope mismatch.
-type HiltSingletonWithActivityDepRule struct {
-	FlatDispatchBase
-	BaseRule
-}
-
-// Confidence reports a tier-2 (medium) base confidence. DI hygiene rule.
-func (r *HiltSingletonWithActivityDepRule) Confidence() float64 { return 0.75 }
-
-var hiltSingletonActivityScopedTypes = map[string]struct{}{
-	"Activity":                  {},
-	"AppCompatActivity":         {},
-	"ComponentActivity":         {},
-	"FragmentActivity":          {},
-	"Fragment":                  {},
-	"DialogFragment":            {},
-	"BottomSheetDialogFragment": {},
-	"View":                      {},
-	"ViewGroup":                 {},
-	"LifecycleOwner":            {},
 }
 
 // SingletonOnMutableClassRule detects @Singleton (or other application-scoped)
